Remove unnecessary casts

+++ src/cortex_var/core/db_variants.c
@@ -762,8 +762,7 @@ double get_log_likelihood_of_genotype_on_variant_called_by_bubblecaller(
     return (double)covg_branch_1 * log(theta_one) 
       - theta_one 
       - log_factorial_uint64_t(covg_branch_1) 
-      +
-      (double)cb2 * log(error_rate_per_base);
+      + cb2 * log(error_rate_per_base);
      
   }
   else if (genotype==hom_other)
@@ -778,7 +777,7 @@ double get_log_likelihood_of_genotype_on_variant_called_by_bubblecaller(
 
     return (double)covg_branch_2 * log(theta_other) - theta_other -
            log_factorial_uint64_t(covg_branch_2) +
-      (double)cb1 * log(error_rate_per_base);
+           cb1 * log(error_rate_per_base);
   }
   else if (genotype==het)
   {
